Try to fix rounding on x86 in _check_date_b_delta
On 32-bit platforms, rounding to an `Int` means rounding to and
`Int32`, and this can fail when using `NanoDate`s because the
values are too large.  Change this to an explicit `Int64` to try
to avoid this.

+++ src/util.jl
@@ -192,8 +192,8 @@ as `Dates.Nanosecond`s.
 function _check_date_b_delta(t::AbstractTrace)
     ismissing(t.evt.time) && error("trace does not have origin time set")
     t.delta < 1e-9 && error("date calculations do not support sampling intervals < 1 ns")
-    b = Dates.Nanosecond(round(Int, starttime(t)*1_000_000_000))
-    delta_in_ns = round(Int, 1_000_000_000*t.delta)
+    b = Dates.Nanosecond(round(Int64, starttime(t)*1_000_000_000))
+    delta_in_ns = round(Int64, 1_000_000_000*t.delta)
     delta = Dates.Nanosecond(delta_in_ns)
     b, delta
 end
